refactor: now page with layout
